backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
python-multipart==0.0.20
numpy==2.2.6
pillow==11.3.0
torch==2.8.0
timm==1.0.19
einops==0.8.1
torchvision==0.23.0
faiss-cpu==1.12.0
requests==2.32.3
tifffile>=2024.0.0
